server/requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
selenium>=3.141.0,<3.142.0
requests>=2.24.0,<2.25.0
pusher>=3.0.0,<3.1.0
webdriver-manager>=3.2.0,<3.3.0
django-filter>=2.1.0,<2.2.0
psycopg2-binary>=2.8.5,<2.9.0
psycopg2>=2.8.5,<2.9.0
django-cors-headers>=2.5.2,<2.6.0